move files under src and clean up eslint setting files.
