--- packages/zoe/.eslintignore
@@ -1,1 +1,0 @@
-bundle-*.js
