--- packages/zoe/.prettierrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-  trailingComma: "all",
-  singleQuote: true
-};
