--- packages/zoe/index.js
@@ -1,1 +1,0 @@
-export { makeZoe } from './zoe';
