+++ packages/zoe/src/contracts/coveredCall.js
