+++ packages/zoe/src/contracts/publicAuction.js
