+++ packages/zoe/src/inviteConfig.js
