+++ packages/zoe/src/isOfferSafe.js
