+++ packages/zoe/test/unitTests/contracts/helpers/test-offerRules.js
@@ -1,6 +1,6 @@
 import { test } from 'tape-promise/tape';
 
-import { makeUnits } from '../../../../contracts/helpers/offerRules';
+import { makeUnits } from '../../../../src/contracts/helpers/offerRules';
 import { setup } from '../../setupBasicMints';
 
 test('makeUnits', t => {
