+++ packages/zoe/test/unitTests/contracts/test-myFirstDapp.js
@@ -2,10 +2,10 @@ import { test } from 'tape-promise/tape';
 import harden from '@agoric/harden';
 import bundleSource from '@agoric/bundle-source';
 
-import { makeZoe } from '../../../zoe';
+import { makeZoe } from '../../../src/zoe';
 import { setup } from '../setupBasicMints';
 
-const myFirstDappRoot = `${__dirname}/../../../contracts/myFirstDapp`;
+const myFirstDappRoot = `${__dirname}/../../../src/contracts/myFirstDapp`;
 
 test('myFirstDapp with valid offers', async t => {
   try {
