+++ packages/zoe/test/unitTests/contracts/test-simpleExchange.js
@@ -2,10 +2,10 @@ import { test } from 'tape-promise/tape';
 import harden from '@agoric/harden';
 import bundleSource from '@agoric/bundle-source';
 
-import { makeZoe } from '../../../zoe';
+import { makeZoe } from '../../../src/zoe';
 import { setup } from '../setupBasicMints';
 
-const simpleExchangeRoot = `${__dirname}/../../../contracts/simpleExchange`;
+const simpleExchangeRoot = `${__dirname}/../../../src/contracts/simpleExchange`;
 
 test('zoe - simpleExchange', async t => {
   try {
