+++ packages/zoe/test/unitTests/test-areRightsConserved.js
@@ -1,6 +1,6 @@
 import { test } from 'tape-promise/tape';
 
-import { areRightsConserved, transpose } from '../../areRightsConserved';
+import { areRightsConserved, transpose } from '../../src/areRightsConserved';
 import { setup } from './setupBasicMints';
 
 const makeUnitMatrix = (unitOps, extentMatrix) =>
